requirements.txt
pandas
python-dotenv
scikit-learn
numpy
logging
matplotlib
plotly